+++ src/gt/gameentity/Sizable.java
@@ -1,5 +1,5 @@
 package gt.gameentity;
 
 public interface Sizable {
-    void setSize(int width, int height);
+    void setSize(double width, double height);
 }
