+++ src/gt/gamestate/NullGameState.java
@@ -23,7 +23,7 @@ public class NullGameState implements GameState {
     }
 
     @Override
-    public void setSize(int width, int height) {
+    public void setSize(double width, double height) {
         // Do nothing
     }
